frontend: Extract Layout from App
