+++ frontend/src/components/App/AppContainer.tsx
@@ -1,55 +1,13 @@
-import { Box } from '@material-ui/core';
-import AppBar from '@material-ui/core/AppBar';
-import Container from '@material-ui/core/Container';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import Link from '@material-ui/core/Link';
-import { makeStyles } from '@material-ui/core/styles';
-import Toolbar from '@material-ui/core/Toolbar';
 import { SnackbarProvider } from 'notistack';
 import React from 'react';
 import { BrowserRouter, HashRouter } from 'react-router-dom';
 import helpers from '../../helpers';
 import { useTypedSelector } from '../../redux/reducers/reducers';
-import { ClusterTitle } from '../cluster/Chooser';
-import ActionsNotifier from '../common/ActionsNotifier';
-import AlertNotification from '../common/AlertNotification';
 import ReleaseNotes from '../common/ReleaseNotes/ReleaseNotes';
-import Sidebar, { drawerWidth, NavigationTabs } from '../Sidebar';
-import RouteSwitcher from './RouteSwitcher';
-import TopBar from './TopBar';
-
-const useStyle = makeStyles(theme => ({
-  root: {
-    background: theme.palette.background.default,
-    paddingLeft: (props: { isSidebarOpen: boolean }) =>
-      props.isSidebarOpen ? `${drawerWidth}px` : '0px',
-    marginLeft: drawerWidth,
-    '& > *': {
-      color: theme.palette.text.primary,
-    },
-  },
-  content: {
-    flexGrow: 1,
-  },
-  toolbar: theme.mixins.toolbar,
-  // importing visuallyHidden has typing issues at time of writing.
-  // import { visuallyHidden } from '@material-ui/utils';
-  visuallyHidden: {
-    border: 0,
-    clip: 'rect(0 0 0 0)',
-    height: '1px',
-    margin: -1,
-    overflow: 'hidden',
-    padding: 0,
-    position: 'absolute',
-    whiteSpace: 'nowrap',
-    width: '1px',
-  },
-}));
+import Layout from './Layout';
 
 export default function AppContainer() {
   const isSidebarOpen = useTypedSelector(state => state.ui.sidebar.isSidebarOpen);
-  const classes = useStyle({ isSidebarOpen });
   const Router = ({ children }: React.PropsWithChildren<{}>) =>
     helpers.isElectron() ? (
       <HashRouter>{children}</HashRouter>
@@ -67,39 +25,7 @@ export default function AppContainer() {
       }}
     >
       <Router>
-        <Link href="#main" className={classes.visuallyHidden}>
-          Skip to main content
-        </Link>
-        <Box display="flex">
-          <CssBaseline />
-
-          <AppBar
-            position="fixed"
-            className={classes.root}
-            elevation={1}
-            component="nav"
-            aria-label="Appbar Tools"
-          >
-            <Toolbar>
-              <div style={{ flex: '1 0 0' }} />
-              <ClusterTitle />
-              <div style={{ flex: '1 0 0' }} />
-              <TopBar />
-            </Toolbar>
-          </AppBar>
-          <Sidebar />
-          <main id="main" className={classes.content}>
-            <AlertNotification />
-            <Box p={3}>
-              <div className={classes.toolbar} />
-              <Container maxWidth="lg">
-                <NavigationTabs />
-                <RouteSwitcher />
-              </Container>
-            </Box>
-          </main>
-          <ActionsNotifier />
-        </Box>
+        <Layout isSidebarOpen={isSidebarOpen} />
       </Router>
       <ReleaseNotes />
     </SnackbarProvider>
